requirements.txt
streamlit
slack_sdk
python-dotenv
google-generativeai
requests
